Version 0.0.52
* More tests
