+++ src/test/helpers/purchaseOrder/createPurchaseOrder.specHelper.ts
@@ -1,16 +1,12 @@
 import { expect } from 'vitest';
 
 import { BgNodeClient } from '../../../BgNodeClient.js';
-import chance from '../../../helpers/chance.js';
 import logger from '../../../helpers/logger.js';
 import { PurchaseOrder } from '../../../models/PurchaseOrder.js';
 import { ShoppingCartItem } from '../../../models/ShoppingCartItem.js';
 import {
-  findGiftCardProductsSpecHelper,
-} from '../giftCardProduct/findGiftCardProducts.specHelper.js';
-import {
-  createShoppingCartItemSpecHelper,
-} from '../shoppingCartItem/createShoppingCartItem.specHelper.js';
+  createMultipleShoppingCartItemsSpecHelper,
+} from '../shoppingCartItem/createMultipleShoppingCartItems.specHelper.js';
 
 export const createPurchaseOrderSpecHelper = async (
   props: Partial<PurchaseOrder> | undefined,
@@ -18,8 +14,11 @@ export const createPurchaseOrderSpecHelper = async (
   shoppingCartItems: ShoppingCartItem[],
   client: BgNodeClient,
 ): Promise<{ purchaseOrder: PurchaseOrder, shoppingCartItems: ShoppingCartItem[] }> => {
-  logger.debug('BgServiceApiCheck.createPurchaseOrder: calling API/createPurchaseOrder',
-    { props });
+  logger.debug('createPurchaseOrderSpecHelper called.', { props, itemCount });
+
+  if (shoppingCartItems.length < 1) {
+    shoppingCartItems = await createMultipleShoppingCartItemsSpecHelper(itemCount, client);
+  }
 
   if (!props) {
     props = {};
@@ -35,34 +34,6 @@ export const createPurchaseOrderSpecHelper = async (
 
   if (!props.createdBy) {
     props.createdBy = client.clientInfoStore.myUserId;
-  }
-
-  if (shoppingCartItems.length < 1) {
-    const products = await findGiftCardProductsSpecHelper(
-      undefined,
-      { limit: 20 },
-      client,
-    );
-    const productsWithDenominations = products.filter(p => p.denominations?.length > 0);
-    const selectedProducts = chance.pickset(productsWithDenominations, itemCount);
-
-    for (const product of selectedProducts) {
-      const itemProps: Partial<ShoppingCartItem> = {
-        shoppingCartId: client.clientInfoStore.myUserId,
-        createdBy: client.clientInfoStore.myUserId,
-        productId: product.id,
-        quantity: chance.integer({ min: 1, max: Math.min(3, productsWithDenominations.length) }),
-      };
-      const denomination = chance.pickone(product.denominations);
-      itemProps.price = denomination?.amount || 10000; // $100.00
-      itemProps.totalPrice = itemProps.price * (itemProps.quantity || 1);
-      const item = await createShoppingCartItemSpecHelper(itemProps, client);
-
-      if (!item) {
-        throw new Error('Failed to create shopping cart item');
-      }
-      shoppingCartItems.push(item);
-    }
   }
 
   const response = await client.operations.purchaseOrder.createPurchaseOrder(props);
