+++ src/test/operations/shoppingCart/findMyShoppingCart.spec.ts
@@ -2,8 +2,12 @@ import { afterAll, beforeAll, describe, expect, test } from 'vitest';
 
 import { BgNodeClient } from '../../../BgNodeClient.js';
 import { CachePolicy } from '../../../enums.js';
+import chance from '../../../helpers/chance.js';
 import { MyUser } from '../../../models/MyUser.js';
 import clientStore from '../../helpers/clientStore.js';
+import {
+  createMultipleShoppingCartItemsSpecHelper
+} from '../../helpers/shoppingCartItem/createMultipleShoppingCartItems.specHelper.js';
 import { deleteMyUserSpecHelper } from '../../helpers/user/deleteMyUser.specHelper.js';
 import { signMeUpSpecHelper } from '../../helpers/user/signMeUp.specHelper.js';
 
@@ -32,7 +36,10 @@ describe('operations.shoppingCart.findMyShoppingCart', () => {
   //   expect(cachedShoppingCart.items.length).toBe(0);
   // });
 
-  test('should return the cached user from the local db', async () => {
+  test('returns the shopping cart from the network', async () => {
+    const itemCount = chance.integer({ min: 2, max: 6 });
+    const shoppingCartItems = await createMultipleShoppingCartItemsSpecHelper(itemCount, client);
+
     const networkResult = await client.operations.shoppingCart.findMyShoppingCart({
       cachePolicy: CachePolicy.network,
     });
@@ -41,6 +48,6 @@ describe('operations.shoppingCart.findMyShoppingCart', () => {
     expect(networkResult.error).toBeUndefined();
     expect(networkResult.object).toBeDefined();
     expect(shoppingCart.id).toBe(myUser.id);
-    expect(shoppingCart.items.length).toBe(0);
+    expect(shoppingCart.items.length).toBe(itemCount);
   });
 });
